Removed mainapplication class
