+++ src/main/java/org/jabref/gui/JabRefGUI.java
@@ -3,6 +3,7 @@ package org.jabref.gui;
 import java.util.List;
 import java.util.Optional;
 
+import javafx.application.Application;
 import javafx.application.Platform;
 import javafx.scene.Scene;
 import javafx.scene.input.KeyEvent;
@@ -13,6 +14,7 @@ import javafx.stage.WindowEvent;
 import org.jabref.gui.help.VersionWorker;
 import org.jabref.gui.icon.IconTheme;
 import org.jabref.gui.keyboard.TextInputKeyBindings;
+import org.jabref.gui.openoffice.OOBibBaseConnect;
 import org.jabref.gui.theme.ThemeManager;
 import org.jabref.logic.UiCommand;
 import org.jabref.logic.l10n.Localization;
@@ -21,7 +23,7 @@ import org.jabref.logic.util.WebViewStore;
 import org.jabref.model.strings.StringUtil;
 import org.jabref.model.util.FileUpdateMonitor;
 import org.jabref.preferences.GuiPreferences;
-import org.jabref.preferences.PreferencesService;
+import org.jabref.preferences.JabRefPreferences;
 
 import com.tobiasdiez.easybind.EasyBind;
 import impl.org.controlsfx.skin.DecorationPane;
@@ -31,29 +33,34 @@ import org.slf4j.LoggerFactory;
 /**
  * Represents the outer stage and the scene of the JabRef window.
  */
-public class JabRefGUI {
+public class JabRefGUI extends Application {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(JabRefGUI.class);
 
+    private static List<UiCommand> uiCommands;
+    private static JabRefPreferences preferencesService;
+    private static FileUpdateMonitor fileUpdateMonitor;
     private static JabRefFrame mainFrame;
     private static DialogService dialogService;
     private static ThemeManager themeManager;
 
-    private final Stage mainStage;
-    private final PreferencesService preferencesService;
+    private boolean correctedWindowPos = false;
+    private Stage mainStage;
 
-    private final List<UiCommand> uiCommands;
-    private boolean correctedWindowPos;
+    public static void setup(List<UiCommand> uiCommands,
+                             JabRefPreferences preferencesService,
+                             FileUpdateMonitor fileUpdateMonitor) {
+        JabRefGUI.uiCommands = uiCommands;
+        JabRefGUI.preferencesService = preferencesService;
+        JabRefGUI.fileUpdateMonitor = fileUpdateMonitor;
+    }
 
-    public JabRefGUI(Stage mainStage,
-                     List<UiCommand> uiCommands,
-                     PreferencesService preferencesService,
-                     FileUpdateMonitor fileUpdateMonitor) {
-        this.mainStage = mainStage;
-        this.uiCommands = uiCommands;
-        this.preferencesService = preferencesService;
+    @Override
+    public void start(Stage stage) {
+        this.mainStage = stage;
 
-        this.correctedWindowPos = false;
+        FallbackExceptionHandler.installExceptionHandler();
+        Globals.startBackgroundTasks();
 
         WebViewStore.init();
 
@@ -88,6 +95,13 @@ public class JabRefGUI {
         });
 
         setupProxy();
+    }
+
+    @Override
+    public void stop() {
+        OOBibBaseConnect.closeOfficeConnection();
+        Globals.stopBackgroundTasks();
+        Globals.shutdownThreadPools();
     }
 
     private void setupProxy() {
