+++ src/main/java/org/jabref/gui/desktop/os/DefaultDesktop.java
@@ -5,8 +5,8 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Path;
 
+import org.jabref.Launcher;
 import org.jabref.architecture.AllowedToUseAwt;
-import org.jabref.cli.Launcher;
 import org.jabref.gui.DialogService;
 import org.jabref.preferences.FilePreferences;
