+++ src/main/java/org/jabref/gui/desktop/os/Linux.java
@@ -12,8 +12,8 @@ import java.util.List;
 import java.util.Locale;
 import java.util.Optional;
 
+import org.jabref.Launcher;
 import org.jabref.architecture.AllowedToUseAwt;
-import org.jabref.cli.Launcher;
 import org.jabref.gui.DialogService;
 import org.jabref.gui.JabRefExecutorService;
 import org.jabref.gui.externalfiletype.ExternalFileType;
